Split window and render/update thread

+++ core/src/main/java/org/elder/core/ecs/System.java
@@ -1,5 +1,5 @@
 package org.elder.core.ecs;
 
 public interface System {
-    void update(long delta);
+    void update(float delta);
 }
